move transaction history to components, simplify import statements
